api_with_restrictions/requirements.txt
django
djangorestframework
psycopg2
django-filter
